refactor(frontend): rename vendored liquor-tree to tasktree
Move tasks/assets/liquor-tree/src to tasks/assets/tasktree and drop the
vendored package metadata (README, LICENSE, package.json). After the Vue 3
migration the tree component was rewritten to Composition API and no longer
resembles upstream liquor-tree, so the vendored name was misleading.

Update the Tree import in Board.vue, the two intra-package imports that
reach up into tasks/assets (now one level shallower), and the path
reference in the _context-menu.scss comment.
