--- tasks/assets/liquor-tree/package.json
@@ -1,108 +1,0 @@
-{
-  "_from": "liquor-tree",
-  "_id": "liquor-tree@0.2.67",
-  "_inBundle": false,
-  "_integrity": "sha512-uwzLBkJxGNY3GiBhhByNlLq/HD30sOnuZUTMQCJwO//C8/Wl+zzs4jLk1cjdvppUU+08vPVA6I6W9Fn7IyD3YQ==",
-  "_location": "/liquor-tree",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "tag",
-    "registry": true,
-    "raw": "liquor-tree",
-    "name": "liquor-tree",
-    "escapedName": "liquor-tree",
-    "rawSpec": "",
-    "saveSpec": null,
-    "fetchSpec": "latest"
-  },
-  "_requiredBy": [
-    "#USER",
-    "/"
-  ],
-  "_resolved": "https://registry.npmjs.org/liquor-tree/-/liquor-tree-0.2.67.tgz",
-  "_shasum": "b7f70033399be468e0e6f289645a72494d941f73",
-  "_spec": "liquor-tree",
-  "_where": "/Users/dragonee/Kod/tasks",
-  "author": {
-    "name": "Kostiantyn",
-    "email": "phlyze@gmail.com"
-  },
-  "bugs": {
-    "url": "https://github.com/amsik/liquor-tree/issues"
-  },
-  "bundleDependencies": false,
-  "dependencies": {},
-  "deprecated": false,
-  "description": "A Vue.js tree component.",
-  "devDependencies": {
-    "@storybook/addon-notes": "^3.4.11",
-    "@storybook/addon-options": "^3.4.11",
-    "@storybook/vue": "^3.4.11",
-    "@vue/test-utils": "^1.0.0-beta.25",
-    "babel-core": "^6.26.3",
-    "babel-jest": "^22.4.4",
-    "babel-plugin-transform-builtin-extend": "^1.1.2",
-    "babel-preset-env": "^1.7.0",
-    "cross-env": "^5.2.0",
-    "eslint": "^4.19.1",
-    "eslint-plugin-jest": "^21.22.1",
-    "eslint-plugin-vue-libs": "^2.1.0",
-    "jest": "^22.4.4",
-    "rollup": "^0.66.2",
-    "rollup-plugin-alias": "^1.4.0",
-    "rollup-plugin-buble": "^0.19.2",
-    "rollup-plugin-serve": "^0.4.2",
-    "rollup-plugin-uglify": "^3.0.0",
-    "rollup-plugin-vue": "^3.0.0",
-    "vue": "^2.5.17",
-    "vue-jest": "^2.6.0",
-    "vue-loader": "^14.2.3",
-    "vue-template-compiler": "^2.5.17"
-  },
-  "files": [
-    "src",
-    "dist/*.js"
-  ],
-  "homepage": "https://amsik.github.io/liquor-tree/",
-  "jest": {
-    "setupFiles": [
-      "./test/setupTests.js"
-    ],
-    "moduleFileExtensions": [
-      "js",
-      "vue"
-    ],
-    "moduleNameMapper": {
-      "^@/(.*)$": "<rootDir>/src/$1"
-    },
-    "transform": {
-      "^.+\\.js$": "<rootDir>/node_modules/babel-jest",
-      ".*\\.(vue)$": "<rootDir>/node_modules/vue-jest"
-    },
-    "testEnvironment": "node"
-  },
-  "keywords": [
-    "vue",
-    "tree",
-    "component",
-    "vue-tree",
-    "treeview"
-  ],
-  "library": "LiquorTree",
-  "main": "dist/liquor-tree.umd.js",
-  "module": "dist/liquor-tree.esm.js",
-  "name": "liquor-tree",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/amsik/liquor-tree.git"
-  },
-  "scripts": {
-    "build": "cross-env NODE_ENV=production rollup -c rollup.config.js",
-    "dev": "cross-env NODE_ENV=development rollup -w -c rollup.config.js",
-    "lint": "eslint --fix src test",
-    "storybook": "start-storybook -p 9001 -c ./docs/storybook -s ./docs/storybook/public",
-    "test": "npm run lint && npm run unit",
-    "unit": "jest"
-  },
-  "version": "0.2.67"
-}
