+++ tasks/assets/tasktree/components/DraggableNode.vue
