+++ tasks/assets/tasktree/lib/Node.js
