+++ tasks/assets/tasktree/lib/Tree.js
